+++ PSFramework/changelog.md
@@ -3,6 +3,7 @@
 ## ???
 
 - New: Added Tab Expansion Plus Plus code to provide extended tab completion on PS3-4
+- New: Argument Transformation Attribute: PsfDynamicTransform - allows dynamic object conversion from PSObject, hashtable, or type from a different library version
 - Upd: Invoke-PSFProtectedCommand now respects explicitly bound `-WhatIf` and `-Confirm` parameters.
 
 ## 1.3.140-preview2 : 2020-08-20